# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Quantum Commerce API
  description: API oficial do case fictício Quantum Commerce — disciplina Tools, Automations and Workflows (FIAP). Backend
    determinístico para agentes n8n, Dify, Power Automate e Python. Sem LLM no backend.
  version: 2.0.0
paths:
...
  /api/v1/returns/check-eligibility:
    post:
      tags:
      - Returns
      summary: Check Return Eligibility
      operationId: check_return_eligibility
      parameters:
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnEligibilityRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnEligibilityResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/returns:
    post:
      tags:
      - Returns
      summary: Create Return
      operationId: create_return
      parameters:
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnResponse'
        '200':
          description: Idempotent replay
        '404':
          description: Order not found
        '409':
          description: Idempotency key conflict
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/returns/{return_id}:
    get:
      tags:
      - Returns
      summary: Get Return By Id
      operationId: get_return_by_id
      parameters:
      - name: return_id
        in: path
        required: true
        schema:
          type: string
          title: Return Id
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Return By Id
        '404':
          description: Return not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/customers/{customer_id}/returns:
    get:
      tags:
      - Returns
      summary: List Customer Returns
      operationId: list_customer_returns
      parameters:
      - name: customer_id
        in: path
        required: true
        schema:
          type: string
          title: Customer Id
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response List Customer Returns
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ReturnCreate:
      properties:
        customer_id:
          type: string
          title: Customer Id
        order_id:
          type: string
          title: Order Id
        sku:
          type: string
          title: Sku
        reason:
          type: string
          title: Reason
      type: object
      required:
      - customer_id
      - order_id
      - sku
      - reason
      title: ReturnCreate
    ReturnEligibilityRequest:
      properties:
        order_id:
          type: string
          title: Order Id
        sku:
          type: string
          title: Sku
      type: object
      required:
      - order_id
      - sku
      title: ReturnEligibilityRequest
    ReturnEligibilityResponse:
      properties:
        eligible:
          type: boolean
          title: Eligible
        reason:
          type: string
          title: Reason
        days_since_delivery:
          anyOf:
          - type: integer
          - type: 'null'
          title: Days Since Delivery
        return_window_days:
          type: integer
          title: Return Window Days
        policy_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Policy Id
      type: object
      required:
      - eligible
      - reason
      - return_window_days
      title: ReturnEligibilityResponse
    ReturnResponse:
      properties:
        return_id:
          type: string
          title: Return Id
        status:
          type: string
          title: Status
        protocol:
          type: string
          title: Protocol
      type: object
      required:
      - return_id
      - status
      - protocol
      title: ReturnResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...